requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
arrow==1.1.1
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
ConfigArgParse==1.5.3
configparser==5.0.2
cryptography==35.0.0
DBUtils==2.0.2
Faker==8.12.0
Flask==2.0.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==20.9.0
geventhttpclient==1.5.3
geventhttpclient-wheels==1.3.1.dev2
greenlet==1.1.2
h11==0.12.0
idna==3.2
importlib-metadata==4.6.4
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonpath==0.82
locustio==0.14.6
MarkupSafe==2.0.1
msgpack==1.0.2
numpy==1.21.3
opencv-contrib-python==4.5.4.58
outcome==1.1.0
packaging==21.0
Pillow==8.4.0
pluggy==0.13.1
psutil==5.8.0
py==1.10.0
pycparser==2.20
pycryptodome==3.11.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.4
pytest-assume==2.4.3
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.1
python-dateutil==2.8.2
pywin32==302
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.26.0
roundrobin==0.0.2
ruamel.yaml==0.17.13
ruamel.yaml.clib==0.2.6
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
text-unidecode==1.3
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
